Refactor: Outsources code for closest point query as standalone class in quest
Also renames class to `DistributedClosestPoint`.
