+++ src/axom/quest/CMakeLists.txt
@@ -21,7 +21,8 @@ set( quest_headers
 
     Delaunay.hpp
     SignedDistance.hpp
-
+    DistributedClosestPoint.hpp
+    
     ## All-nearest-neighbors query
     AllNearestNeighbors.hpp
     detail/AllNearestNeighbors_detail.hpp
